# Avoid copy in itoa and write directly to the target buffer