```
$ benchstat old-add-time.log new-add-time.log
name           old time/op    new time/op    delta
LTSVLog-2         712ns ± 1%     675ns ± 0%  -5.19%  (p=0.000 n=10+9)
StandardLog-2     806ns ± 0%     805ns ± 0%    ~     (p=0.163 n=9+10)

name           old alloc/op   new alloc/op   delta
LTSVLog-2          238B ± 0%      238B ± 0%    ~     (all equal)
StandardLog-2      274B ± 0%      274B ± 0%    ~     (all equal)

name           old allocs/op  new allocs/op  delta
LTSVLog-2          3.00 ± 0%      3.00 ± 0%    ~     (all equal)
StandardLog-2      3.00 ± 0%      3.00 ± 0%    ~     (all equal)
```